refactor: organize code into separate files for maintainablitiy
