+++ blocks/to-do/to-do.js
@@ -1,45 +1,29 @@
+import handleKeyPress from '../../scripts/key-press.js';
+import addListItem from '../../scripts/add-list-items.js';
+
+// Selectors
+const toDoAddDiv = document.querySelector('.to-do > div');
+const addBtn = document.querySelector('.to-do .button-container');
+
+// Create element
+const input = document.createElement('input');
+const orderedList = document.createElement('ol');
+
+// Decorate block function
 export default function decorate(block) {
-  // Selectors
-  const toDoAddDiv = document.querySelector('.to-do > div');
-  const addBtn = document.querySelector('.to-do .button-container');
-
-  // Create element
-  const input = document.createElement('input');
-  const orderedList = document.createElement('ol');
-
   // Change content
   input.setAttribute('placeholder', 'Type here...');
   toDoAddDiv.appendChild(input);
   block.appendChild(orderedList);
 
-  // Functions
-  // Add item function
-  const addListItem = (event) => {
-    // Prevent button from scrolling up
-    event.preventDefault();
-
-    const inputValue = input.value.trim();
-
-    // Make sure input is not empty
-    if (inputValue !== '') {
-      const listItem = document.createElement('li');
-      listItem.textContent = inputValue;
-      orderedList.appendChild(listItem);
-
-      // Empty input after
-      input.value = '';
-    }
+  // Specify addListItem function
+  const addToDo = (event) => {
+    addListItem(event, input, orderedList);
   };
 
-  // Handle key press function
-  const handleKeyPress = (event) => {
-    // Check if the Enter key was pressed
-    if (event.key === 'Enter') {
-      addListItem(event);
-    }
-  };
-
-  // Call function
-  addBtn.addEventListener('click', addListItem);
-  input.addEventListener('keypress', handleKeyPress);
+  // Event listeners
+  addBtn.addEventListener('click', addToDo);
+  input.addEventListener('keypress', (event) => {
+    handleKeyPress(event, addToDo, input);
+  });
 }
